requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==3.0.1
django-debug-toolbar==3.2.1
django-heroku==0.3.1
django-mathfilters==1.0.0
dnspython==2.1.0
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pygal==2.4.0
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
wrapt==1.12.1
